fix: wrap Apollo provider on client only
